requirements.txt
asgiref==3.4.1
Django==3.2.13
djangorestframework==3.13.1
Pillow==8.4.0
pkg_resources==0.0.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
